🚚 Move FileListView to /documents/views
