+++ src/documents/views/StudyFilesListView.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {graphql, compose} from 'react-apollo';
-import {GET_STUDY_BY_ID, MY_PROFILE} from '../state/queries';
+import {GET_STUDY_BY_ID, MY_PROFILE} from '../../state/queries';
 import {UploadContainer} from '../containers';
 import FileList from '../components/FileList/FileList';
 import {
@@ -12,7 +12,6 @@ import {
   Segment,
   Button,
 } from 'semantic-ui-react';
-
 import UploadWizard from '../modals/UploadWizard/UploadWizard';
 
 /**
