+++ src/documents/views/index.js
@@ -1,0 +1,3 @@
+export {default as StudyFilesListView} from './StudyFilesListView';
+export {default as FileDetailView} from './FileDetailView';
+export {default as NewDocumentView} from './NewDocumentView';
