+++ src/views/index.js
@@ -1,13 +1,10 @@
 export {default as StudyListView} from './StudyListView';
-export {default as StudyFilesListView} from './StudyFilesListView';
 export {default as LoginView} from './LoginView';
 export {default as CallbackView} from './CallbackView';
-export {default as FileDetailView} from './FileDetailView';
 export {default as NavBarView} from './NavBarView';
 export {default as EmptyView} from './EmptyView';
 export {default as TokensListView} from './TokensListView';
 export {default as CavaticaProjectsView} from './CavaticaProjectsView';
-export {default as NewDocumentView} from './NewDocumentView';
 export {default as NewStudyView} from './NewStudyView';
 export {default as ProfileView} from './ProfileView';
 export {default as StudyInfoView} from './StudyInfoView';
